Fix realtime outbox worker server-only import
